Simplify calling of the Parser in Guides.
Okay, this change is significantly bigger than anticipated .. What I
aimed to accomplish in this change is to create a simple api for the
Parser class that can be used without too much hassle.

The API is now simplified to:

- __construct()
- prepare()
- parse()

See the README.md for usage instructions.

Along the way I came across problematic dependencies, junk methods and
properties and more of such. Many I have marked with new TODO's and some
I have fixed in this change.

After this change, the TODO's need to be resolved and the Renderer needs
to be addressed because it tightly couples with phpDocumentor itself and
this coupling needs to be reduced if we want to publish guides as a
separate library